merge w/ master

+++ Scripts/Services/PointsSystems/CleanUpBritanniaData.cs
@@ -13,7 +13,7 @@ namespace Server.Engines.Points
         public override PointsType Loyalty { get { return PointsType.CleanUpBritannia; } }
         public override TextDefinition Name { get { return m_Name; } }
         public override bool AutoAdd { get { return true; } }
-        public override double MaxPoints { get { return int.MaxValue; } }
+        public override double MaxPoints { get { return double.MaxValue; } }
         public override bool ShowOnLoyaltyGump { get { return false; } }
 
         private TextDefinition m_Name = null;
